Only hit the server to get tags if the tags dialog is open.  Otherwise use "pending tags" in CampaignContact to change display of tags button.  Before we were hitting the server each time contact changed.
